Generalization of code to accodomate user-defined SSCs
This commit adds the ability to add more SSCs into the SSCs folder
on the server. If defined in a similar manner to the predefined
ones, the client will add a selection button and generate the
table based on the definition.

checkDevCompliance also received a small switch to accomodate for
checking of SSC specific functionality which cannot be described
in a JSON (f.e. check of PSID or suggestions of minor versions for
Opal 2).

Fixed typo