Basic auth with token for authorization added
Proper secrets handling etc. will be made later on. For now the
client authenticates with HTTP Basic auth. A signed session token
with lifespan of 8 hours is then sent to the client to authorize
all requests which modify data on server-side. Basic testing was
done over this, but further will be done during user testing phase.