Added timestamps to prevent needless fetching
Server now returns list of drive names along with UTC timestamps
of last modification. Client compares this its local timestamp and
if a drive is out-of-date then it fetches it from the server. This
was done to ease the load on the server.

!Important: If you have old data stored in IndexedDB from before
this commit, then use the debug button to re-fetch the database.